+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.1.1 (Next)
 
 * Your contribution here.
+* [#1](https://github.com/dblock/actionmailer-text/pull/1) Ignore elements outside the body - [@woahdae](https://github.com/woahdae).
 
 ### 0.1.0 (2015-04-07)